requirements.txt
django
djangorestframework
python-dotenv
python-decouple
psycopg2-binary